+++ android/src/main/kotlin/pl/hasoft/better_player/BetterPlayerCache.kt
@@ -1,4 +1,4 @@
-package com.jhomlala.better_player
+package pl.hasoft.better_player
 
 import android.content.Context
 import android.util.Log